added token validation

+++ server/sim-card-manager/src/main/java/com/artezio/simcardmanager/api/Controller.java
@@ -19,6 +19,7 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 
+import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
 
 @RestController
@@ -38,37 +39,66 @@ public class Controller {
     }
 
     @GetMapping("/cards")
-    public Iterable<Card> getAll() {
-        Iterable<Card> all = this.cardRepository.findAll();
-        return all;
+    public Iterable<Card> getAll(HttpServletRequest request) {
+        if (this.checkRequestToken(request)){
+            Iterable<Card> all = this.cardRepository.findAll();
+            return all;
+        } else {
+            throw new NotAuthorizedException();
+        }
     }
 
     @GetMapping("/cards/{id}")
-    public Card getOne(@PathVariable(value = "id") Long id) {
-        Card one = this.cardRepository.findOne(id);
-        return one;
+    public Card getOne(@PathVariable(value = "id") Long id,
+                       HttpServletRequest request) {
+        if (this.checkRequestToken(request)){
+            Card one = this.cardRepository.findOne(id);
+            return one;
+        } else {
+            throw new NotAuthorizedException();
+        }
     }
 
     @DeleteMapping("/cards/{id}")
-    public void deleteOne(@PathVariable(value = "id") Long id) {
-        this.cardRepository.delete(id);
+    public void deleteOne(@PathVariable(value = "id") Long id,
+                          HttpServletRequest request) {
+        if (this.checkRequestToken(request)){
+            this.cardRepository.delete(id);
+        } else {
+            throw new NotAuthorizedException();
+        }
     }
 
     @PostMapping("/cards")
     @ResponseStatus(HttpStatus.CREATED)
-    public Card createOne(@RequestBody Card card) {
-        return this.cardRepository.save(card);
+    public Card createOne(@RequestBody Card card,
+                          HttpServletRequest request) {
+        if (this.checkRequestToken(request)) {
+            return this.cardRepository.save(card);
+        } else {
+            throw new NotAuthorizedException();
+        }
     }
 
     @PutMapping("/cards")
-    public Card updateOne(@RequestBody Card card) {
-        return this.cardRepository.save(card);
+    public Card updateOne(@RequestBody Card card,
+                          HttpServletRequest request) {
+        if (this.checkRequestToken(request)){
+            return this.cardRepository.save(card);
+        } else {
+            throw new NotAuthorizedException();
+        }
     }
 
     @GetMapping("/cards/search")
-    public Iterable<Card> search(@RequestParam(value = "number") String number) {
-        Iterable<Card> byNumberLike = this.cardRepository.findByNumberContaining(number);
-        return byNumberLike;
+    public Iterable<Card> search(@RequestParam(value = "number") String number,
+                                 HttpServletRequest request) {
+        if (this.checkRequestToken(request)) {
+            Iterable<Card> byNumberLike = this.cardRepository.findByNumberContaining(number);
+            return byNumberLike;
+        } else {
+            throw new NotAuthorizedException();
+        }
     }
 
     @PostMapping("/login")
@@ -92,6 +122,15 @@ public class Controller {
     @GetMapping("/register")
     public Boolean checkRegister(@RequestParam (value = "login") String login) {
         return this.userRepository.findByLogin(login) == null;
+    }
+
+    //this is secure service for generate and check JWT
+    private boolean checkRequestToken(HttpServletRequest request) {
+        boolean result = false;
+        if (request.getHeader("x-auth-token") != null) {
+            result = request.getHeader("x-auth-token").equals(this.generateSecureJwt());
+        }
+        return result;
     }
 
     private String generateSecureJwt() {
